+++ Source/AGXUnreal/Public/Utilities/AGX_ROS2Utilities.h
@@ -16,10 +16,10 @@ class AGXUNREAL_API FAGX_ROS2Utilities
 {
 public:
 	static FAGX_SensorMsgsImage Convert(
-		const TArray<FColor>& Image, float TimeStamp, const FIntPoint& Resolution, bool Grayscale);
+		const TArray<FColor>& Image, double TimeStamp, const FIntPoint& Resolution, bool Grayscale);
 
 	static FAGX_SensorMsgsImage Convert(
-		const TArray<FFloat16Color>& Image, float TimeStamp, const FIntPoint& Resolution,
+		const TArray<FFloat16Color>& Image, double TimeStamp, const FIntPoint& Resolution,
 		bool Grayscale);
 };
