+++ Source/AGXUnrealBarrier/Private/SimulationBarrier.cpp
@@ -223,16 +223,16 @@ bool FSimulationBarrier::GetEnableAMOR()
 	return NativeRef->Native->getMergeSplitHandler()->getEnable();
 }
 
-void FSimulationBarrier::SetTimeStep(float TimeStep)
+void FSimulationBarrier::SetTimeStep(double TimeStep)
 {
 	check(HasNative());
 	NativeRef->Native->setTimeStep(ConvertToAGX(TimeStep));
 }
 
-float FSimulationBarrier::GetTimeStep() const
+double FSimulationBarrier::GetTimeStep() const
 {
 	check(HasNative());
-	return ConvertToUnreal<float>(NativeRef->Native->getTimeStep());
+	return ConvertToUnreal<double>(NativeRef->Native->getTimeStep());
 }
 
 void FSimulationBarrier::SetEnableContactWarmstarting(bool bEnable)
@@ -449,13 +449,13 @@ void FSimulationBarrier::Step()
 	}
 }
 
-float FSimulationBarrier::GetTimeStamp() const
+double FSimulationBarrier::GetTimeStamp() const
 {
 	check(HasNative());
-	return ConvertToUnreal<float>(NativeRef->Native->getTimeStamp());
+	return ConvertToUnreal<double>(NativeRef->Native->getTimeStamp());
 }
 
-void FSimulationBarrier::SetTimeStamp(float TimeStamp)
+void FSimulationBarrier::SetTimeStamp(double TimeStamp)
 {
 	check(HasNative());
 	NativeRef->Native->setTimeStamp(ConvertToAGX(TimeStamp));
